Tried my best to copy
I tried to create a mobile version first, but i'm struggling with a few things 1. The space between the two divs (can I just make one div and that's it?) 2. I used negative margins to close down that space which I know is not the right thing to do so I have to figure out something else